requirements.txt
opencv-python
deepface
tf_keras
django
channels
daphne
websockets
webrtcvad-wheels
pyaudio
openai
groq
requests
